backend/requirements.txt
flask-marshmallow==0.14.0
flask-restplus==0.13.0
flask-sqlalchemy==2.4.4
flask-migrate==3.1.0
PyMySQL==1.0.2
flask==1.1.2
jinja2==2.11.2
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
sqlalchemy==1.3.20
werkzeug==0.16.1
apifairy==0.9.0
Flask-Cors==3.0.10
marshmallow_enum==1.5.1
requests==2.27.1
python-dotenv==0.19.2
pandas==1.1.5
openpyxl==3.0.9
markupsafe==2.0.1
itsdangerous==2.0.1